3rd commit
게시글 수정,삭제 구현
메인 페이지 디자인
// 1. 게시글 내용 한글 깨짐.
    2. 게시글 삭제시 글번호 정렬 추가하기.

+++ view.jsp
@@ -9,7 +9,7 @@
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="viewport" content="width=device-width" , initial-scale="1">
 <link rel="stylesheet" href="css/bootstrap.css">
-
+<link rel="stylesheet" href="css/custom.css">
 <title>JSP 게시판 웹사이트</title>
 </head>
 <body>
@@ -101,7 +101,7 @@
 						</tr>
 						<tr>
 							<td>내용</td>
-							<td colspan="2" style ="min-height:200px; text-align: Left;"><%=bbs.getBbsContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">","&rt") %></td>
+							<td colspan="2" style="min-height:200px; text-align: Left;"><%=bbs.getBbsContent().replaceAll(" ", "&nbsp;").replaceAll("<", "&lt;").replaceAll(">","&gt;").replaceAll("\n", "<br>") %></td>
 						</tr>
 					</tbody>
 				</table>
@@ -110,7 +110,7 @@
 					if(userID != null && userID.equals(bbs.getUserID())){
 				%>
 						<a href="update.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">수정</a>
-						<a href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
+						<a onclick=" return confirm('정말 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
 				<%
 					}
 				%>
